double-check that get_a_logprobs is correct

+++ atari_irl/irl.py
@@ -322,8 +322,8 @@ class AtariAIRL(AIRL):
         # build energy model
         with tf.variable_scope(name) as _vs:
             # Should be batch_size x T x dO/dU
-            self.obs_t = tf.placeholder(tf.float32, list((None,) + self.dOshape), name='obs')
-            self.nobs_t = tf.placeholder(tf.float32, list((None,) + self.dOshape), name='nobs')
+            self.obs_t = tf.placeholder(tf.int32, list((None,) + self.dOshape), name='obs')
+            self.nobs_t = tf.placeholder(tf.int32, list((None,) + self.dOshape), name='nobs')
             self.act_t = tf.placeholder(tf.float32, [None, self.dU], name='act')
             self.nact_t = tf.placeholder(tf.float32, [None, self.dU], name='nact')
             self.labels = tf.placeholder(tf.float32, [None, 1], name='labels')
